pyproject.toml
[tool.poetry]
name = "atomic-habits"
version = "0.1.0"
description = ""
authors = ["Ramazan-Z <Ramazan_zalilov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
redis = "^5.2.1"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ipython = "^8.30.0"
coverage = "^7.6.9"

[tool.poetry.group.lint.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.13.0"
django-stubs = "^5.1.1"
djangorestframework-stubs = "^3.15.1"
types-psycopg2 = "^2.9.21.20241019"
django-filter-stubs = "^0.1.3"
types-requests = "^2.32.0.20241016"
drf-yasg-stubs = "^0.1.4"
celery-stubs = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = false
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | incutils
  )/
)
'''
